pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.1",
    "numpy>=2.2.5",
    "openai>=1.78.1",
    "pdfplumber>=0.11.6",
    "python-dotenv>=1.1.0",
    "supabase>=2.15.1",
    "tqdm>=4.67.1",
]
